# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WeTravel API
  description: Python backend with Clerk sessions and Supabase. Authorize with a session
    JWT, never a secret key.
  version: 2.0.0
paths:
...
  /api/posts:
    get:
      tags:
      - Posts
      summary: List Posts
      description: Optional exact country filter; 20 posts per page.
      operationId: list_posts_api_posts_get
      security:
      - ClerkSession: []
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: country
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Country
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_PublishedPost_'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
  /api/posts/{post_id}:
    get:
      tags:
      - Posts
      summary: Get Post
      operationId: get_post_api_posts__post_id__get
      security:
      - ClerkSession: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishedPost'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
    delete:
      tags:
      - Posts
      summary: Delete Post
      operationId: delete_post_api_posts__post_id__delete
      security:
      - ClerkSession: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Post Id
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
  /api/posts/{post_id}/reuse:
    post:
      tags:
      - Posts
      summary: Reuse Post
      description: Create a new plan with reset bookings and empty journals, preserving
        attribution.
      operationId: reuse_post_api_posts__post_id__reuse_post
      security:
      - ClerkSession: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Post Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReuseInput'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneyResponse'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
  /api/posts/{post_id}/save:
    put:
      tags:
      - Posts
      summary: Save Post
      operationId: save_post_api_posts__post_id__save_put
      security:
      - ClerkSession: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Post Id
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
    delete:
      tags:
      - Posts
      summary: Unsave Post
      operationId: unsave_post_api_posts__post_id__save_delete
      security:
      - ClerkSession: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Post Id
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
  /api/posts/{post_id}/like:
    put:
      tags:
      - Posts
      summary: Like Post
      operationId: like_post_api_posts__post_id__like_put
      security:
      - ClerkSession: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Post Id
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
    delete:
      tags:
      - Posts
      summary: Unlike Post
      operationId: unlike_post_api_posts__post_id__like_delete
      security:
      - ClerkSession: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-zA-Z0-9_-]+$
          title: Post Id
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Request Entity Too Large
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Too Many Requests
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
...
components:
  schemas:
    BookingStatus:
      type: string
      enum:
      - Not booked
      - Booked
      - No booking needed
      - Cancelled
      title: BookingStatus
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
        issues:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Issues
      type: object
      required:
      - error
      title: ErrorResponse
...
    JourneyResponse:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        destination:
          type: string
          title: Destination
        startDate:
          type: string
          title: Startdate
        endDate:
          type: string
          title: Enddate
        status:
          $ref: '#/components/schemas/JourneyStatus'
        stops:
          items:
            $ref: '#/components/schemas/StopInput'
          type: array
          title: Stops
        source:
          anyOf:
          - $ref: '#/components/schemas/SourceResponse'
          - type: 'null'
        publishedId:
          anyOf:
          - type: string
          - type: 'null'
          title: Publishedid
      type: object
      required:
      - id
      - title
      - destination
      - startDate
      - endDate
      - status
      - stops
      title: JourneyResponse
    JourneyStatus:
      type: string
      enum:
      - Planning
      - Traveling
      - Completed
      title: JourneyStatus
...
    Page_PublishedPost_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PublishedPost'
          type: array
          title: Items
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - items
      - offset
      - limit
      title: Page[PublishedPost]
...
    PublishedPost:
      properties:
        id:
          type: string
          title: Id
        user_id:
          type: string
          title: User Id
        author_name:
          type: string
          title: Author Name
        title:
          type: string
          title: Title
        destination:
          type: string
          title: Destination
        country:
          type: string
          title: Country
        caption:
          type: string
          title: Caption
        duration_days:
          type: integer
          title: Duration Days
        cover_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Cover Path
        source_post_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Post Id
        source_author:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Author
        source_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Title
        published_at:
          type: string
          format: date-time
          title: Published At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        published_stops:
          items:
            $ref: '#/components/schemas/PublishedStop'
          type: array
          title: Published Stops
      type: object
      required:
      - id
      - user_id
      - author_name
      - title
      - destination
      - country
      - caption
      - duration_days
      - cover_path
      - source_post_id
      - source_author
      - source_title
      - published_at
      - updated_at
      - published_stops
      title: PublishedPost
    PublishedStop:
      properties:
        id:
          type: string
          title: Id
        trip_id:
          type: string
          title: Trip Id
        kind:
          $ref: '#/components/schemas/StopKind'
        name:
          type: string
          title: Name
        day_number:
          type: integer
          title: Day Number
        end_day_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: End Day Number
        review:
          type: string
          title: Review
        rating:
          type: integer
          title: Rating
        photo_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Path
        sort_order:
          type: integer
          title: Sort Order
      type: object
      required:
      - id
      - trip_id
      - kind
      - name
      - day_number
      - end_day_number
      - review
      - rating
      - photo_path
      - sort_order
      title: PublishedStop
...
    ReuseInput:
      properties:
        startDate:
          type: string
          title: Startdate
          examples:
          - '2026-08-01'
      additionalProperties: false
      type: object
      required:
      - startDate
      title: ReuseInput
...
    SourceResponse:
      properties:
        id:
          type: string
          title: Id
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      type: object
      required:
      - id
      title: SourceResponse
    StopInput:
      properties:
        id:
          type: string
          maxLength: 128
          minLength: 1
          pattern: ^[a-zA-Z0-9_-]+$
          title: Id
        kind:
          $ref: '#/components/schemas/StopKind'
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        day:
          type: integer
          maximum: 365.0
          minimum: 0.0
          title: Day
        endDay:
          anyOf:
          - type: integer
            maximum: 365.0
            minimum: 1.0
          - type: 'null'
          title: Endday
        booking:
          $ref: '#/components/schemas/BookingStatus'
          default: Not booked
        visited:
          type: boolean
          title: Visited
          default: false
        confirmation:
          type: string
          maxLength: 1000
          title: Confirmation
          default: ''
        bookingLink:
          type: string
          title: Bookinglink
          default: ''
        cost:
          type: string
          pattern: ^(?:\d{1,10}(?:\.\d{1,2})?)?$
          title: Cost
          default: ''
        currency:
          type: string
          pattern: ^[A-Z]{3}$
          title: Currency
          default: USD
        cancellationDate:
          type: string
          title: Cancellationdate
          default: ''
        review:
          type: string
          maxLength: 2000
          title: Review
          default: ''
        rating:
          type: integer
          maximum: 5.0
          minimum: 0.0
          title: Rating
          default: 0
        photoUri:
          anyOf:
          - type: string
            maxLength: 400
          - type: 'null'
          title: Photouri
      additionalProperties: false
      type: object
      required:
      - id
      - kind
      - name
      - day
      title: StopInput
    StopKind:
      type: string
      enum:
      - Hotel
      - Restaurant
      - Activity
      title: StopKind
...
  securitySchemes:
    ClerkSession:
      type: http
      description: Clerk session JWT from getToken(); never your secret key.
      scheme: bearer